avoid uninitialised QProcess usage
`mount.readAllStandardError()` yields the empty string as the QProcess is not
set up yet, i.e. mounting a disk via clicking on it in Dolphin does not work
and instead just shows an unhelpful error.

Remove the block that appears further down in the function again.
Mounting still fails, but with a proper error this time.

Needs more fixing, but this part is obviously wrong.